feat!: rewrite entire application in go
BREAKING CHANGE: The entire application is now written in Go. This makes it possible to run the proxy with an existing envoy installation and decouples it from Nix. The tools also gained a full POSIC-compatible CLI interface with proper help for all available options. You can now easily download pre-built binaries of the proxy or use Go to install it on your machine.